Fix compiler warnings

+++ libscpi/inc/scpi/parser.h
@@ -61,8 +61,8 @@ extern "C" {
     void SCPI_InitHeap(scpi_t * context, char * error_info_heap, size_t error_info_heap_length);
 #endif
 
-    SCPI_EXPORT scpi_bool_t SCPI_Input(scpi_t* context, const char* data, int len);
-    SCPI_EXPORT scpi_bool_t SCPI_Parse(scpi_t* context, const char* data, int len);
+    SCPI_EXPORT scpi_bool_t SCPI_Input(scpi_t* context, const char* data, size_t len);
+    SCPI_EXPORT scpi_bool_t SCPI_Parse(scpi_t* context, const char* data, size_t len);
 
     SCPI_EXPORT size_t SCPI_ResultCharacters(scpi_t* context, const char* data, size_t len);
 #define SCPI_ResultMnemonic(context, data) SCPI_ResultCharacters((context), (data), strlen(data))
